Move ExtendedStringLiteralTestClass.java
